+++ lib/components/Form/index.ts
@@ -1,0 +1,10 @@
+import Form from "./Form";
+
+
+export type { FormProps, FormHeaderProps, FormBodyProps, FormFooterProps } from "./FormProps";
+
+export { FormHeader as RtFormHeader } from "./Form";
+export { FormBody as RtFormBody } from "./Form";
+export { FormFooter as RtFormFooter } from "./Form";
+
+export default Form;