+++ lib/components/Modal/index.ts
@@ -1,0 +1,4 @@
+import Modal from "./Modal";
+
+export type {ModalProps} from './ModalProps'
+export default Modal;