+++ lib/components/Tabs/index.ts
@@ -1,0 +1,4 @@
+// import TabPane from "./TabPane";
+
+export type {TabPaneProps} from "./TabPane";
+export {default as TabPane} from "./TabPane";