+++ lib/components/TreeSelect/index.ts
@@ -1,0 +1,4 @@
+import TreeSelect from "./TreeSelect";
+
+export type {TreeSelectProps} from './TreeSelectProps'
+export default TreeSelect;