+++ lib/index.tsx
@@ -1,0 +1,11 @@
+// export * from "./importsjs";
+
+export {deprecated} from './components/deprecated'
+// Store
+export {default as rtdReducer} from './redux/rtd.reducer';
+export {setDataStore} from './redux/rtd.actions';
+// Utils
+export {notificationError} from './components/utils/baseUtils'
+export {executeRequest} from "./components/utils/api";
+
+export * from "./components/index";