+++ src/redux/rtd.actions.js
@@ -1,11 +1,6 @@
 import {actionTypes} from './rtd.action.types';
 
-export const initStore = (path) => ({
-    type: actionTypes.INIT_STORE,
-    payload: {path}
-});
-
-export const setDateStore = (path, row) => ({
+export const setDataStore = (path, row) => ({
     type: actionTypes.SET_DATA_STORE,
     payload: {path, row}
 });